+++ src/main/java/edu/woodson/Utility.java
@@ -1,4 +1,4 @@
-package edu.woodoson;
+package edu.woodson;
 
 import java.util.Random;
