+++ src/main/java/edu/woodson/structure/node/Node.java
@@ -4,7 +4,7 @@ import java.util.Arrays;
 import java.util.HashSet;
 import java.util.Set;
 
-import static edu.woodoson.Utility.random;
+import static edu.woodson.Utility.random;
 
 /**
  * @author SirMathhman
